+++ Binaries/Win64/UnrealEditor.modules
@@ -2,6 +2,6 @@
 	"BuildId": "27405482",
 	"Modules": 
 	{
-		"Project_LLL": "UnrealEditor-Project_LLL-0112.dll"
+		"Project_LLL": "UnrealEditor-Project_LLL-5258.dll"
 	}
 }
