--- Source/Project_LLL/Private/GAS/Ability/Object/Interactive/Base/LLL_InteractiveObjectGameplayAbilityBase.cpp
@@ -1,4 +1,0 @@
-﻿// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "GAS/Ability/Object/Interactive/Base/LLL_InteractiveObjectGameplayAbilityBase.h"
