--- Source/Project_LLL/Private/GAS/Ability/Object/Interactive/Reward/LLL_RGA_GiveAbility.cpp
@@ -1,14 +1,0 @@
-﻿// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "GAS/Ability/Object/Interactive/Reward/LLL_RGA_GiveAbility.h"
-
-#include "AbilitySystemBlueprintLibrary.h"
-#include "AbilitySystemComponent.h"
-
-void ULLL_RGA_GiveAbility::ActivateAbility(const FGameplayAbilitySpecHandle Handle, const FGameplayAbilityActorInfo* ActorInfo, const FGameplayAbilityActivationInfo ActivationInfo, const FGameplayEventData* TriggerEventData)
-{
-	Super::ActivateAbility(Handle, ActorInfo, ActivationInfo, TriggerEventData);
-
-	UE_LOG(LogTemp, Log, TEXT("능력 부여"))
-}
