--- Source/Project_LLL/Public/GAS/Ability/Object/Interactive/Base/LLL_InteractiveObjectGameplayAbilityBase.h
@@ -1,16 +1,0 @@
-﻿// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "GAS/Ability/Object/Base/LLL_ObjectGameplayAbilityBase.h"
-#include "LLL_InteractiveObjectGameplayAbilityBase.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class PROJECT_LLL_API ULLL_InteractiveObjectGameplayAbilityBase : public ULLL_ObjectGameplayAbilityBase
-{
-	GENERATED_BODY()
-};
