+++ Source/Project_LLL/Public/Game/LLL_AbilityManageSubSystem.h
@@ -35,16 +35,13 @@ public:
 	virtual void Initialize(FSubsystemCollectionBase& Collection) override;
 	virtual void Deinitialize() override;
 
-	FORCEINLINE TArray<TSoftClassPtr<ULLL_ExtendedGameplayEffect>> GetPlayerGameplayEffects() { return PlayerGameplayEffects; }
-	FORCEINLINE TArray<TSoftClassPtr<ULLL_ExtendedGameplayEffect>> GetMonsterGameplayEffects() { return MonsterGameplayEffects; }
-	FORCEINLINE TArray<TSoftClassPtr<ULLL_ExtendedGameplayEffect>> GetObjectGameplayEffects() { return ObjectGameplayEffects; }
-	FORCEINLINE TArray<TSoftClassPtr<ULLL_ExtendedGameplayEffect>> GetShareableGameplayEffects() { return ShareableGameplayEffects; }
-
 	// 어빌리티 OR 스탯 부여 이펙트 비동기 로딩
 public:
 	void ASyncLoadEffectsByTag(FAsyncLoadEffectDelegate Delegate, EEffectOwnerType Owner, const FGameplayTagContainer& EffectTag, bool TagHasMatching = false);
 	void ASyncLoadAttributeEffectsByTag(FAsyncLoadEffectDelegate Delegate, EEffectOwnerType Owner, const FGameplayTagContainer& EffectTag, bool TagHasMatching = false);
 	void ASyncLoadGrantAbilityEffectsByTag(FAsyncLoadEffectDelegate Delegate, EEffectOwnerType Owner, const FGameplayTagContainer& EffectTag, bool TagHasMatching = false);
+
+	void ASyncLoadEffectsByID(FAsyncLoadEffectDelegate Delegate, EEffectOwnerType Owner, int32 ID, bool TagHasMatching = false);
 	
 private:
 	/** 플레이어에게 게임 진행 도중 일시적으로 어빌리티 OR 스탯을 부여하기 위해 사용하는 이펙트 모음. ex: 보상 \n
@@ -75,5 +72,6 @@ private:
 private:
 	void LoadEffectsFromPath(TArray<TSoftClassPtr<ULLL_ExtendedGameplayEffect>>& Container, FName PrimaryTypes);
 
+public:
 	TArray<TSoftClassPtr<ULLL_ExtendedGameplayEffect>>& GetDataSetByOwner(EEffectOwnerType Owner);
 };
